Refactor model_builder into package
